+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Adjust Anvil panels for clarity
 - Significantly(!) improve hotspot performance
 - Fix edge selection edge case (2 overlapping front faces) 
+- Hotspots can now be used without an external file
+- The hotspots external file path can be chosen
 
 ## 1.4.7
 - Shift-G toggles between snap modes